+++ src/main/java/com/programmers/heycake/common/exception/CustomAuthenticationEntryPoint.java
@@ -1,0 +1,47 @@
+package com.programmers.heycake.common.exception;
+
+import static org.springframework.http.MediaType.*;
+
+import java.io.IOException;
+import java.io.PrintWriter;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.AuthenticationEntryPoint;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.SerializationFeature;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import com.programmers.heycake.common.dto.ErrorResponse;
+
+public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
+
+	@Override
+	public void commence(HttpServletRequest request, HttpServletResponse response,
+			AuthenticationException authException) throws IOException {
+
+		ErrorResponse errorResponse = ErrorResponse.of(
+				"사용자 인증에 실패하였습니다.",
+				request.getRequestURI(),
+				null
+		);
+
+		response.setCharacterEncoding("UTF-8");
+		response.setContentType(APPLICATION_JSON.toString());
+		response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
+
+		PrintWriter writer = response.getWriter();
+
+		ObjectMapper objectMapper = new ObjectMapper();
+		objectMapper.registerModule(new JavaTimeModule());
+		objectMapper.disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+		objectMapper.writeValueAsString(errorResponse);
+
+		writer.println(errorResponse);
+		writer.flush();
+		writer.close();
+	}
+}
+
