--- src/main/java/com/programmers/heycake/common/jwt/CustomAuthenticationEntryPoint.java
@@ -1,33 +1,0 @@
-package com.programmers.heycake.common.jwt;
-
-import java.io.IOException;
-import java.io.OutputStream;
-import java.util.Collections;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import org.springframework.http.MediaType;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.programmers.heycake.common.dto.ErrorResponse;
-
-public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
-
-	@Override
-	public void commence(HttpServletRequest request, HttpServletResponse response,
-			AuthenticationException authException) throws IOException {
-
-		ErrorResponse errorResponse = ErrorResponse.of("사용자 인증에 실패하였습니다.", "", Collections.emptyList());
-		response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-		response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-
-		OutputStream outputStream = response.getOutputStream();
-		ObjectMapper objectMapper = new ObjectMapper();
-		objectMapper.writeValue(outputStream, errorResponse);
-		outputStream.flush();
-	}
-}
-
